scale_trino_workers should *always* be wrapped in a @task when inside a DAG (#223)
Because this function wasn't wrapped in a @task decorator, it was being executed every 30 seconds (!!!) when the Airflow scheduler read the DAG. Then a few seconds later, the Airflow scheduler would scale the cluster down to 1 worker. And so on.

Compare the usage [here](https://github.com/konzainc/konza-dags/blob/main/dags/C_111/dim_accid_to_mpi.py#L23). 

Crucially, we use the `@task` decorator to wrap the function, which means its execution is deferred to run-time, rather than DAG compile-time.

The absence of this decorator was causing the following behavior:
- there would constantly be 20 trino worker nodes up
- every 30 seconds or so, 20 trino worker pods would cycle in and out of existence (thus keeping the trino worker nodes up and preventing the cluster autoscaler from scaling the nodegroup down)
- any trino queries running on these trino workers would fail, since the workers hosting them were being scaled down very quickly.